chore: Refactor refs, and delete package-lock.json because yarn.lock already exists and it is huge as a file
