+++ .gitignore
@@ -111,3 +111,4 @@ tmp/
 temp/
 
 # End of https://www.gitignore.io/api/node
+package-lock.json
