+++ src/revive-rehype.ts
@@ -5,7 +5,7 @@ import { handler as code } from './plugins/code';
 import { hast as figure } from './plugins/figure';
 import { hast as math } from './plugins/math';
 import { handler as ruby } from './plugins/ruby';
-import { inspect } from './utils/debug';
+import { inspect } from './utils';
 
 export default [
   [
